Aumentando y eliminando una vacante al registrarse y utilizando la libreria date-fns

+++ controllers/registro.js
@@ -1,4 +1,8 @@
+// Date-fns
+const { format } = require('date-fns');
+
 // Models
+const Horario = require('../models/horario');
 const Registro = require('../models/registro');
 
 // Obtener Registros
@@ -34,7 +38,7 @@ const crearRegistro = async (request, response) => {
     })
 
     // Valores del body
-    const { email } = response.status;
+    const { email, dni, horario } = request.body;
 
     try {
         // TODO: Encontrar el email en los registros creados
@@ -47,8 +51,68 @@ const crearRegistro = async (request, response) => {
             })
         }
 
+        const existeDNI = await Registro.findOne({ dni });
+
+        // Existe dni
+        if (existeDNI) {
+            return response.status(400).json({
+                ok: false,
+                msg: 'Hay un usuario registrado con ese dni'
+            })
+        }
+
+        // TODO: Buscar horario por el ID 
+        const horarioID = await Horario.findById(horario); 
+
+        // Si no se encuentra el id
+        if (!horarioID) {
+            return response.status(404).json({
+                ok: false,
+                msg: 'No existe el horario por ese id'
+            })
+        }
+
+        // Si no hay vacantes
+        if (horarioID.cant_matriculas <= 0) {
+            return response.status(404).json({
+                ok: false,
+                msg: 'No hay vacantes para este horario'
+            })
+        }
+
+        // TODO: Fechas horario
+        // Fecha inicio
+        let yearS = new Date(horarioID.fecha_inicial).getUTCFullYear();
+        let monthS = new Date(horarioID.fecha_inicial).getUTCMonth();
+        let dayS = new Date(horarioID.fecha_inicial).getUTCDate();
+        // date-fns
+        let fecha_inicio = format(new Date(yearS, monthS, dayS), 'yyyy-MM-dd');
+
+         // Fecha hoy
+        let year = new Date().getFullYear();
+        let month = new Date().getMonth();
+        let day = new Date().getDate();
+        // date-fns
+        let today = format(new Date(year, month, day), 'yyyy-MM-dd');
+
+        // Si la fecha inicial del horario ya pasó
+        if (fecha_inicio < today) {
+            return response.status(404).json({
+                ok: false,
+                msg: 'El horario ya ha comenzado',
+            })
+        }
+
+        // TODO: Reducir una matrícula al horario
+        campo = {
+            cant_matriculas: horarioID.cant_matriculas - 1
+        }
+
         // TODO: Guardar en la base de datos
         const registroDB = await registro.save();
+
+        // TODO: Actualización del horario
+        await Horario.findByIdAndUpdate(horario, campo, { new: true });
 
         response.json({
             ok: true,
@@ -133,6 +197,25 @@ const eliminarRegistro = async (request, response) => {
                 msg: 'No existe un registro por ese id'
             })
         }
+
+        // TODO: Buscar horario por el ID 
+        const horarioID = await Horario.findById(registroID.horario); 
+
+        // Si no se encuentra el id
+        if (!horarioID) {
+            return response.status(404).json({
+                ok: false,
+                msg: 'No existe el horario por ese id'
+            })
+        }
+
+        // TODO: Aumentar una matrícula al horario
+        campo = {
+            cant_matriculas: horarioID.cant_matriculas + 1
+        }
+
+        // TODO: Actualización del horario
+        await Horario.findByIdAndUpdate(registroID.horario, campo, { new: true });
 
         // TODO: Eliminación
         await Registro.findByIdAndDelete(uid);
